(ref): Refactored files to be in its own module and expose its own services

+++ src/modules/notes/notes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common'
+import { Injectable, InternalServerErrorException, NotFoundException, UnauthorizedException } from '@nestjs/common'
 import { CreateNoteDto } from './dto/create-note.dto'
 import { calculateReadTimeMinutes, createEmptyRatings, extractBodyFromFile } from './helper'
 import { NotesRepository } from './notes.repository'
@@ -6,12 +6,13 @@ import { Note } from './models/notes.model'
 import { IConfigAttributes } from '../../common/interfaces/config/app-config.interface'
 import { getConfig } from '../../config'
 import { existsSync } from 'fs'
+import { FilesService } from '../files/files.service'
 
 const config: IConfigAttributes = getConfig()
 
 @Injectable()
 export class NotesService {
-	constructor(private readonly notesRepository: NotesRepository) {}
+	constructor(private readonly notesRepository: NotesRepository, private readonly filesService: FilesService) {}
 
 	async getNoteWithId(id: number): Promise<Note> {
 		const note: Note = await this.notesRepository.findNoteById(id)
@@ -65,7 +66,7 @@ export class NotesService {
 		return this.notesRepository.updateNote(note, filteredData)
 	}
 
-	async deleteNoteWithId(authorId: number, id: number): Promise<boolean> {
+	async deleteNoteWithId(authorId: number, id: number, type: string): Promise<boolean> {
 		const note: Note = await this.notesRepository.findNoteById(id)
 
 		if (!note) {
@@ -74,6 +75,15 @@ export class NotesService {
 
 		if (authorId !== note.authorId) {
 			throw new UnauthorizedException('You are not allowed to delete this note as you are not its author')
+		}
+
+		// Delete the actual file for the note
+		const delSuccess = await this.filesService.deleteFileWithId({ id: note.fileId, type: type })
+
+		if (!delSuccess) {
+			throw new InternalServerErrorException(
+				'For some reason, we could not delete the file associated with this note. Aborting delete.'
+			)
 		}
 
 		return this.notesRepository.deleteNote(note)
